feat: convert project to typescript
- Rename all .js files to .ts
- Add and configure tsconfig.json
- Update vite configs for typescript
- Update manifest.json for background script
- Add typescript and type definitions
- Fix type errors
- Setup yarn pnp for vscode
