+++ package.json
@@ -22,9 +22,13 @@
   },
   "homepage": "https://github.com/da2ce7/deepseek-export#readme",
   "devDependencies": {
+    "@types/chrome": "^0.1.1",
+    "@types/dompurify": "^3.2.0",
+    "@types/node": "^24.1.0",
     "eslint": "^9.31.0",
     "globals": "^16.3.0",
     "prettier": "^3.6.2",
+    "typescript": "^5.8.3",
     "vite": "^7.0.5",
     "vite-plugin-static-copy": "^3.1.1"
   },
